Factor our the function to create organizations
